Added a table of contents, to keep track of where the story is (roughly).

+++ app/src/main/java/compx576/assignment/httydgame/npcDAO.java
@@ -15,7 +15,7 @@ public interface npcDAO {
     NPC findNPCByName(String name);
 
     @Query("UPDATE characters SET relLevel = :updatedValue WHERE name = :characterName")
-    void updateRelationship(String characterName, int updatedValue);
+    void updateRelationship(String characterName, float updatedValue);
 
     @Query("SELECT * FROM characters")
     List<NPC> getAllNPCs();
